refactored code_analyzer
