+++ services/code_analyzer/code_snippet.py
@@ -8,10 +8,3 @@ class CodeSnippet:
     tree: ast.Module
     start_line: int
     end_line: int
-
-    def to_json(self):
-        return {
-            "code": self.code,
-            "start_line": self.start_line,
-            "end_line": self.end_line
-        }
