+++ services/code_analyzer/requirements.txt
@@ -3,16 +3,22 @@ certifi==2023.11.17
 charset-normalizer==3.3.2
 click==8.1.7
 colorama==0.4.6
+exceptiongroup==1.2.0
 Flask==3.0.0
 Flask-Cors==4.0.0
 idna==3.6
+iniconfig==2.0.0
 itsdangerous==2.1.2
 Jinja2==3.1.2
 mando==0.7.1
 MarkupSafe==2.1.3
+packaging==23.2
+pluggy==1.3.0
 Pygments==2.17.2
-radon==6.0.1
+pytest==7.4.3
+python-dotenv==1.0.0
 requests==2.31.0
 six==1.16.0
+tomli==2.0.1
 urllib3==2.1.0
 Werkzeug==3.0.1
