+++ services/db_interface/model/__init__.py
@@ -14,9 +14,6 @@ from .user_rating_type import UserRatingType  # noqa: F401
 from .code_snippet import CodeSnippet  # noqa: F401
 from .prompt_part_usage import PromptPartUsage  # noqa: F401
 from .stop_sequence_usage import StopSequenceUsage  # noqa: F401
-from .symbol_definition import SymbolDefinition  # noqa: F401
-from .symbol_definition_type import SymbolDefinitionType  # noqa: F401
 from .symbol_reference import SymbolReference  # noqa: F401
-from .symbol_definition_reference import SymbolDefinitionReference  # noqa: F401, E501
-from .undefined_symbol_reference import UndefinedSymbolReference  # noqa: 401
+from .symbol_reference import SymbolReference  # noqa: 401
 from .system_prompt import SystemPrompt  # noqa: 401
