+++ services/db_interface/model/symbol_reference.py
@@ -1,32 +1,17 @@
 from model.base_model import BaseModel
-from peewee import IntegerField
-from typing import List
+from model.code_snippet import CodeSnippet
+from peewee import TextField, IntegerField, ForeignKeyField
 
 
 class SymbolReference(BaseModel):
     class Meta:
         table_name = "SymbolReference"
 
+    symbol = TextField()
     start_line = IntegerField()
     end_line = IntegerField()
     start_column = IntegerField()
     end_column = IntegerField()
-
-    @property
-    def symbol_definition_references(self) -> List[BaseModel]:
-        from model import SymbolDefinitionReference
-        return self.get_backref_list_1_n(
-            SymbolDefinitionReference,
-            SymbolDefinitionReference.symbol_reference,
-            self
-        )
-
-    @property
-    def symbol_definitions(self) -> List[BaseModel]:
-        from model import SymbolDefinition, SymbolDefinitionReference
-        return self.get_backref_list_m_n(
-            SymbolDefinition,
-            SymbolDefinitionReference,
-            SymbolDefinitionReference.symbol_reference,
-            self
-        )
+    code_snippet = ForeignKeyField(
+        CodeSnippet, backref='_symbol_references'
+    )
