+++ tests/tests_00_db_interface/test_02_post.py
@@ -34,11 +34,7 @@ class TestDatabaseInterfacePost:
         "frameworks",
         "prompt_part_usages",
         "stop_sequence_usages",
-        "symbol_definitions",
-        "symbol_definition_types",
         "symbol_references",
-        "symbol_definition_references",
-        "undefined_symbol_references",
         "user_rating_types"
     ])
     def test_post_instance(
