+++ tests/util/dummy_data.json
@@ -2,9 +2,6 @@
     "follow_up_types": {
         "name": "test_follow_up_types"
     },
-    "symbol_definition_types": {
-        "name": "test_symbol_definition_types"
-    },
     "user_rating_types": {
         "name": "test_user_rating_types"
     },
@@ -79,55 +76,14 @@
         "code": "This is a code snippet.",
         "start_line": 1,
         "end_line": 2,
-        "raw_loc": 3,
-        "raw_lloc": 4,
-        "raw_sloc": 5,
-        "raw_comments": 6,
-        "raw_multi": 7,
-        "raw_single_comments": 8,
-        "raw_blank": 9,
-        "halstead_h1": 10,
-        "halstead_h2": 11,
-        "halstead_N1": 12,
-        "halstead_N2": 13,
-        "halstead_length": 14.0,
-        "halstead_volume": 15.0,
-        "halstead_difficulty": 16.0,
-        "halstead_effort": 17.0,
-        "halstead_time": 18.0,
-        "halstead_bugs": 19.0,
-        "cyclomatic_complexity_score": 20.0,
-        "cyclomatic_complexity_rank": "A",
-        "maintainability_index_score": 21.0,
-        "maintainability_index_rank": "A",
         "prediction": "$predictions"
     },
-    "undefined_symbol_references": {
+    "symbol_references": {
         "symbol": "This is an undefined symbol reference.",
         "start_line": 1,
         "end_line": 2,
         "start_column": 3,
         "end_column": 4,
         "code_snippet": "$code_snippets"
-    },
-    "symbol_definitions": {
-        "symbol": "This is a symbol definition.",
-        "start_line": 1,
-        "end_line": 2,
-        "start_column": 3,
-        "end_column": 4,
-        "is_builtin": true,
-        "code_snippet": "$code_snippets",
-        "symbol_definition_type": "$symbol_definition_types"
-    },
-    "symbol_references": {
-        "start_line": 1,
-        "end_line": 2,
-        "start_column": 3,
-        "end_column": 4
-    },
-    "symbol_definition_references": {
-        "symbol_definition": "$symbol_definitions",
-        "symbol_reference": "$symbol_references"
     }
 }
